Isolate correspondence collision fixtures
Move the compiler-produced identity collision cases out of the shared C# diff corpus so its locked smoke baseline remains unchanged.


Copilot-Session: c7ddf0f0-3d68-4a8c-b2b6-86b8ecd5c24d
