+++ src/DiffFixtures.V1/DiffSample.cs
@@ -1,9 +1,5 @@
 using System.Collections.Generic;
 
-public class T
-{
-}
-
 namespace DiffFixtureSample
 {
     public static class DiffSample
@@ -478,27 +474,5 @@ namespace DiffFixtureSample
     public class InternalExplicitSurface : IInternalExplicitSurface
     {
         int IInternalExplicitSurface.Get() => 1;
-    }
-}
-
-namespace CorrespondenceIdentity
-{
-    public class GenericCollisions
-    {
-        public int NamedType<T>(
-            T value,
-            global::T other) => 1;
-
-        public int PrimitiveName<@int>(@int value) => 1;
-
-        public int Tuple((int left, int right) value) => 1;
-    }
-
-    public class Outer
-    {
-        public class Inner
-        {
-            public int M() => 1;
-        }
     }
 }
