+++ src/DiffFixtures.V2/DiffSample.cs
@@ -1,9 +1,5 @@
 using System.Collections.Generic;
 
-public class T
-{
-}
-
 namespace DiffFixtureSample
 {
     public static class DiffSample
@@ -492,27 +488,5 @@ namespace DiffFixtureSample
     public class InternalExplicitSurface : IInternalExplicitSurface
     {
         int IInternalExplicitSurface.Get() => 2;
-    }
-}
-
-namespace CorrespondenceIdentity
-{
-    public class GenericCollisions
-    {
-        public int NamedType<U>(
-            U value,
-            global::T other) => 2;
-
-        public int PrimitiveName<T>(T value) => 2;
-
-        public int Tuple((int x, int y) value) => 2;
-    }
-}
-
-namespace CorrespondenceIdentity.Outer
-{
-    public class Inner
-    {
-        public int M() => 2;
     }
 }
