+++ src/ILInspector.Research.Tests/ILInspector.Research.Tests.csproj
@@ -25,6 +25,8 @@
     <ProjectReference Include="..\DiffFixtures.V1\DiffFixtures.V1.csproj" ReferenceOutputAssembly="false" />
     <ProjectReference Include="..\DiffFixtures.V2\DiffFixtures.V2.csproj" ReferenceOutputAssembly="false" />
     <ProjectReference Include="..\ILInspector.Research.TargetFixtures\ILInspector.Research.TargetFixtures.csproj" ReferenceOutputAssembly="false" />
+    <ProjectReference Include="..\ResearchTargetCorrespondenceFixtures.V1\ResearchTargetCorrespondenceFixtures.V1.csproj" ReferenceOutputAssembly="false" />
+    <ProjectReference Include="..\ResearchTargetCorrespondenceFixtures.V2\ResearchTargetCorrespondenceFixtures.V2.csproj" ReferenceOutputAssembly="false" />
     <ProjectReference Include="..\ILInspector.Analysis.Fixtures\ILInspector.Analysis.Fixtures.csproj" />
     <ProjectReference Include="..\ILInspector.Decompiler\ILInspector.Decompiler.csproj" />
     <ProjectReference Include="..\ILInspector.ILDiff\ILInspector.ILDiff.csproj" />
